Overwrite stdout.write before creating pino instance
