+++ app/api/server/v1/misc.js
@@ -14,7 +14,7 @@ import { settings } from '../../../settings/server';
 import { API } from '../api';
 import { getDefaultUserFields } from '../../../utils/server/functions/getDefaultUserFields';
 import { getURL } from '../../../utils/lib/getURL';
-import { getQueue } from '../../../../server/stream/logQueue';
+import { getLogs } from '../../../../server/stream/stdout';
 import { SystemLogger } from '../../../../server/lib/logger/system';
 
 
@@ -227,7 +227,7 @@ API.v1.addRoute('stdout.queue', { authRequired: true }, {
 		if (!hasPermission(this.userId, 'view-logs')) {
 			return API.v1.unauthorized();
 		}
-		return API.v1.success({ queue: getQueue() });
+		return API.v1.success({ queue: getLogs() });
 	},
 });
